Refactor bifrost_benchmarks as new module
- Sigproc pipeline benchmark file now creates a class which inherits from the PipelineBenchmarker class. This provides similar functionality for timing future pipelines
