+++ test/benchmarks/bifrost_benchmarks/__init__.py
@@ -1,0 +1,4 @@
+""" Initiate the python module
+"""
+
+from bifrost_benchmarks.pipeline_benchmarker import PipelineBenchmarker
